Add files via upload
Change evdrything when it comes to Fractal Architecture. Instead of these hard lines think more like a nesting bird.  In celestial terms ' Think like a nesting angel. An angel is a bird to begin with. There is a place in the cosmos where angels actualy gather ; forget about all the myths and bible stories. The universe has expanded and so the human must now morph into an a real live angel to become an interplanetary and universal citizen.citezen. Think only butterflies can do metamorphosis ? Fractal architecture is constructed in another dimention other than this one. The "fractal dimention " !